Moved TableColumn to api package
because I feel like it defines the table API instead of being an internal implementation.
It is intended to be extended by zircon users
